fix(jobs): bound unified worker concurrency, wire enqueue notify, fix unconditional trusted_system auth
