+++ crates/axon-core/src/config/parse/toml_config.rs
@@ -372,6 +372,8 @@ pub(super) struct TomlWorkersSection {
     pub ingest_lanes: Option<usize>,
     /// Parallel embed worker lanes.
     pub embed_lanes: Option<usize>,
+    /// Maximum concurrent jobs the unified worker runs at once.
+    pub unified_worker_concurrency: Option<usize>,
     /// Per-document embed timeout in seconds.
     pub embed_doc_timeout_secs: Option<u64>,
     /// Queue summary interval in seconds.
