+++ crates/axon-core/src/config/parse/tuning.rs
@@ -183,6 +183,7 @@ pub(super) fn apply_env_toml_tuning(cfg: &mut Config, toml: &TomlConfig) {
     );
     cfg.ingest_lanes = ingest_lanes(toml);
     cfg.embed_lanes = embed_lanes(toml);
+    cfg.unified_worker_concurrency = unified_worker_concurrency(toml);
     cfg.embed_doc_timeout_secs = embed_doc_timeout_secs(toml);
     cfg.queue_summary_secs = queue_summary_secs(toml);
     cfg.freshness_tick_secs = freshness_tick_secs(toml);
@@ -939,6 +940,16 @@ fn ingest_lanes(toml: &TomlConfig) -> usize {
 
 fn embed_lanes(toml: &TomlConfig) -> usize {
     resolve_clamped_usize("AXON_EMBED_LANES", toml.workers.embed_lanes, 2, 1, 32)
+}
+
+fn unified_worker_concurrency(toml: &TomlConfig) -> usize {
+    resolve_clamped_usize(
+        "AXON_UNIFIED_WORKER_CONCURRENCY",
+        toml.workers.unified_worker_concurrency,
+        8,
+        1,
+        64,
+    )
 }
 
 fn embed_doc_timeout_secs(toml: &TomlConfig) -> u64 {
